add README.md for chatroom
