+++ examples/environments/envs/chatroom.py
@@ -21,6 +21,8 @@ from agentscope.environment import (
     event_func,
 )
 from agentscope.models import ModelResponse
+from agentscope.studio._client import _studio_client
+from agentscope.web.gradio.utils import user_input
 
 
 class ChatRoomMember(BasicEnv):
@@ -55,7 +57,7 @@ class ChatRoomMember(BasicEnv):
         """Make the agent chatting in the chatroom."""
         time.sleep(delay)
         while True:
-            msg = self._agent(Msg(name="user", content="", role="user"))
+            msg = self._agent()
             if 'goodbye' in msg.content.lower():
                 break
             sleep_time = random.randint(1, 5)
@@ -291,4 +293,65 @@ class ChatRoomAgent(AgentBase):
         msg = Msg(name=self.name, content=response, role="assistant")
         if response:
             self.speak(msg)
+        return msg
+
+
+class ChatRoomAgentWithAssistant(ChatRoomAgent):
+    """A ChatRoomAgent with assistant"""
+    def __init__(
+        self,
+        timeout: float | None = None,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.timeout = timeout
+
+    def reply(self, x: Msg = None) -> Msg:
+        if _studio_client.active:
+            logger.info(
+                f"Waiting for input from:\n\n"
+                f"    * {_studio_client.get_run_detail_page_url()}\n",
+            )
+            raw_input = _studio_client.get_user_input(
+                agent_id=self.agent_id,
+                name=self.name,
+                require_url=False,
+                required_keys=None,
+                timeout=self.timeout,
+            )
+
+            logger.info("Python: receive ", raw_input)
+            if raw_input is None:
+                content = None
+            else:
+                content = raw_input["content"]
+        else:
+            time.sleep(0.5)
+            try:
+                content = user_input(timeout=self.timeout)
+            except TimeoutError:
+                content = None
+
+        if content is not None:  # user input
+            response = content
+        else:  # assistant reply
+            msg_hint = self.generate_hint()
+            self_msg = Msg(name=self.name, content=f"", role="assistant")
+
+            history = self.room.get_history(self.agent_id)
+            prompt = self.model.format(
+                msg_hint,
+                history,
+                self_msg,
+            )
+            logger.debug(prompt)
+            response = self.model(
+                prompt,
+                parse_func=self.room.chatting_parse_func,
+                max_retries=3,
+            ).text
+            if not response.startswith('[auto reply]'):
+                response = '[auto reply] ' + response
+        msg = Msg(name=self.name, content=response, role="user")
+        self.speak(msg)
         return msg
